more files that should be part of last commit.

+++ codebase/apps/didss/src/FmqTest/Params.hh
@@ -385,7 +385,7 @@ public:
 
   int output_n_slots;
 
-  int output_buf_size;
+  long output_buf_size;
 
   int data_mapper_reg_interval;
